feat: add filter

+++ cli/src/commands/tui/events/network.rs
@@ -163,7 +163,7 @@ async fn update_logs_entries(app: Arc<Mutex<App>>, log_entries: Option<Vec<LogEn
                 app_ref.state.logs_vertical_scroll_state = app_ref
                     .state
                     .logs_vertical_scroll_state
-                    .position(app_ref.state.log_entries_length);
+                    .position(app_ref.state.log_entries_length as u16);
             }
         }
     }
